fix: 인수테스트 validation과 응답을 수정한다
