+++ src/main/java/shopping/dto/ProductResponse.java
@@ -1,0 +1,66 @@
+package shopping.dto;
+
+
+import java.util.Objects;
+
+public final class ProductResponse {
+
+    private Long id;
+    private String name;
+    private String imageUrl;
+    private String price;
+
+    public ProductResponse() {
+    }
+
+    public ProductResponse(final Long id, final String name, final String imageUrl, final String price) {
+        this.id = id;
+        this.name = name;
+        this.imageUrl = imageUrl;
+        this.price = price;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getImageUrl() {
+        return imageUrl;
+    }
+
+    public String getPrice() {
+        return price;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof ProductResponse)) {
+            return false;
+        }
+        ProductResponse that = (ProductResponse) o;
+        return Objects.equals(id, that.id) && Objects.equals(name, that.name)
+                && Objects.equals(imageUrl, that.imageUrl) && Objects.equals(price, that.price);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, name, imageUrl, price);
+    }
+
+    @Override
+    public String toString() {
+        return "ProductResponse{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", imageUrl='" + imageUrl + '\'' +
+                ", price='" + price + '\'' +
+                '}';
+    }
+}
