+++ src/test/java/shopping/accept/UrlHelper.java
@@ -3,15 +3,14 @@ package shopping.accept;
 import static io.restassured.RestAssured.given;
 
 import io.restassured.response.ValidatableResponse;
-import org.springframework.http.MediaType;
 
 class UrlHelper {
 
     static final class Product {
 
         public static ValidatableResponse getAllProducts() {
-            return given().log().all().accept(MediaType.TEXT_HTML_VALUE)
-                    .when().log().all().get("/products")
+            return given().log().all()
+                    .when().get("/")
                     .then().log().all();
         }
